Better to specify source files instead of using GLOB option. Here the reasons:     https://stackoverflow.com/questions/1027247/is-it-better-to-specify-source-files-with-glob-or-each-file-individually-in-cmak#1060061
Also it is not recommended in the cmake documentation:
    https://cmake.org/cmake/help/v3.4/command/file.html
Note: We do not recommend using GLOB to collect a list of source files from your source tree. If no CMakeLists.txt file changes when a source is added or removed then the generated build system cannot know when to ask CMake to regenerate.